Silence a deprecation warning
I'm not sure why this just now started cropping up but it's an easy
enough fix.

I believe we do want to enforce available locales--if someone is trying
to use the site in an unsupported language, an error page isn't great UX
but neither is just giving them a middle finger, and at least with the
error we get an ExceptionNotification.